+++ aim/web/ui/src/components/ChartPanel/ChartGrid/ChartGrid.d.ts
@@ -15,4 +15,5 @@ export interface IChartGridProps {
   readOnly?: boolean;
   syncHoverState?: (args: ISyncHoverStateArgs) => void;
   resizeMode?: ResizeModeEnum;
+  onMount?: () => void;
 }