+++ aim/web/ui/src/config/pageTitles/pageTitles.ts
@@ -12,6 +12,7 @@ const pageTitles = {
   RUN_DETAIL: '',
   EXPERIMENT: '',
   FIGURES_EXPLORER_NEW: 'New Figures',
+  METRICS_EXPLORER_V2: 'Metrics Explorer v2',
 };
 
 export default pageTitles;