+++ aim/web/ui/src/modules/BaseExplorer/components/Controls/Aggregation/index.ts
@@ -1,0 +1,5 @@
+import Aggregation from './Aggregation';
+
+export * from './Aggregation.d';
+
+export default Aggregation;