+++ aim/web/ui/src/modules/BaseExplorer/components/Controls/Legends/Legends.d.ts
@@ -1,0 +1,5 @@
+import { IBaseComponentProps } from 'modules/BaseExplorer/types';
+
+export interface ILegendsProps extends IBaseComponentProps {
+  visualizationName: string;
+}