+++ aim/web/ui/src/modules/core/pipeline/grouping/getGroupValues.ts
@@ -1,4 +1,4 @@
-import { orderBy } from 'lodash-es';
+import _ from 'lodash-es';
 
 import { buildObjectHash } from 'modules/core/utils/hashing';
 
@@ -30,13 +30,17 @@ function getGroups(
   // generate possible groups
   const groups: Record<string, Group> = data.reduce(
     (groups: Group, value: any) => {
-      const groupValue: GroupValue = pickValues(value, fields);
+      const pickedValues = pickValues(value, fields);
+      if (Object.keys(pickedValues).length === 0) {
+        return groups;
+      }
+      const groupValue: GroupValue = { ...pickedValues, type };
       const groupKey: string = buildObjectHash(groupValue);
 
       if (!groups.hasOwnProperty(groupKey)) {
         groups[groupKey] = {
           key: groupKey,
-          fields: groupValue,
+          fields: _.omit(groupValue, 'type'),
           items: [],
           type,
         };
@@ -51,7 +55,7 @@ function getGroups(
 
   const groupsList = Object.values(groups);
 
-  const orderedGroups = orderBy(
+  const orderedGroups = _.orderBy(
     groupsList,
     fields.map((f) => (group) => getValue(group, ['fields', f])),
     orders,