+++ aim/web/ui/src/routes/routes.tsx
@@ -49,6 +49,9 @@ const FiguresExplore = React.lazy(
 const AudiosExplorer = React.lazy(
   () => import(/* webpackChunkName: "audios" */ 'pages/AudiosExplorer'),
 );
+const MetricsExplorer = React.lazy(
+  () => import(/* webpackChunkName: "metrics_v2" */ 'pages/MetricsExplorer'),
+);
 
 export interface IRoute {
   path: PathEnum;
@@ -200,6 +203,15 @@ const routes = {
     displayName: 'Experiment',
     isExact: false,
     title: pageTitlesEnum.EXPERIMENT,
+  },
+  METRICS_EXPLORER: {
+    path: PathEnum.Metrics_Explorer,
+    component: MetricsExplorer,
+    showInSidebar: false,
+    icon: 'metrics',
+    displayName: 'Metrics_v2',
+    isExact: true,
+    title: pageTitlesEnum.METRICS_EXPLORER_V2,
   },
 };