+++ aim/web/ui/src/types/components/LineChart/LineChart.d.ts
@@ -20,8 +20,6 @@ import { IRun } from 'types/services/models/metrics/runModel';
 
 import { CurveEnum, HighlightEnum } from 'utils/d3';
 
-import { IChartPanelProps } from '../ChartPanel/ChartPanel';
-
 export interface ILine {
   key: string;
   data: {
@@ -36,13 +34,13 @@ export interface ILine {
 }
 
 export interface ILineChartProps {
-  index: number;
-  data: IChartPanelProps['data'];
+  index?: number;
+  id?: string;
+  data: ILine[];
   nameKey?: string;
   aggregatedData?: IAggregatedData[];
   alignmentConfig?: IAlignmentConfig;
   ignoreOutliers?: boolean;
-  axesScaleType: IAxesScaleState;
   axesScaleRange?: IAxesScaleRange;
   highlightMode?: HighlightEnum;
   curveInterpolation?: CurveEnum;
@@ -51,8 +49,22 @@ export interface ILineChartProps {
   chartTitle?: IChartTitle;
   zoom?: IChartZoom;
   onZoomChange?: (zoom: Partial<IChartZoom>) => void;
-  readOnly?: boolean;
   resizeMode?: ResizeModeEnum;
+  onMount?: () => void;
+  axesScaleType?: IAxesScaleState;
+  readOnly?: boolean;
+  margin?: { top: number; right: number; bottom: number; left: number };
+}
+
+export interface ILineChartRef {
+  setActiveLineAndCircle: (
+    lineKey: string,
+    focusedStateActive: boolean = false,
+    force: boolean = false,
+  ) => void;
+  updateHoverAttributes: (xValue: number, dataSelector?: string) => void;
+  clearHoverAttributes: () => void;
+  setFocusedState: (focusedState: IFocusedState) => void;
 }
 
 export interface IUpdateFocusedChartArgs {