+++ aim/web/ui/src/types/utils/d3/drawArea.d.ts
@@ -3,7 +3,8 @@ import React from 'react';
 import { IChartTitle } from 'types/services/models/metrics/metricsAppModel';
 
 export interface IDrawAreaArgs {
-  index: number;
+  index?: number;
+  id: string;
   nameKey: string;
   parentRef: React.MutableRefObject<>;
   visAreaRef: React.MutableRefObject<>;