+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements
 - Expose `run_name` and `run_hash` parameters to the `aim.sdk.adapters.pytorch_lightning.AimLogger` adapter (constd)
+- Implement Metrics explorer v2 via Base explorer (KaroMourad)
 
 ### Fixes